Plant Disease Image Classifier/web app/requirements.txt
streamlit
tensorflow==2.20.0
keras==3.13.2
pillow
numpy